Add new public PWA config file and icons
The commit introduces a new file, 'pwa.json', which is the configuration file for the Progressive Web App (PWA) version of the Car Dealer application. Multiple icons along with their various dimensions have been added in the SVG format for both fullscreen and minimal-ui displays. Screenshots of the application's home and car detail pages, and shortcuts for adding and managing cars, are also included.